Cleanup Alewxa command router Create buildCommandResponseAsync in func-alexa Change logic so that validation comes before response generation (save compute cycles) Add further code comments to command function to ease maint.
