+++ src/routes/default.js
@@ -763,6 +763,15 @@ router.post('/api/v1/devices', defaultLimiter,
 	}
 );
 ///////////////////////////////////////////////////////////////////////////
+// Set State API (Not in Use)
+///////////////////////////////////////////////////////////////////////////
+// router.post('/setstate/:dev_id',
+// 	passport.authenticate(['bearer', 'basic'], { session: false }),
+// 	async (req, res) => {
+// 		// do nothing, disused for now, may use along side command API
+// 	}
+// );
+///////////////////////////////////////////////////////////////////////////
 // Functions
 ///////////////////////////////////////////////////////////////////////////
 function ensureAuthenticated(req,res,next) {
